fix: hide advanced sync settings
