+++ src/components/SettingsAdvancedTransfers.tsx
@@ -1,0 +1,96 @@
+import { InfoCard } from './InfoCard'
+import { LabeledValueRow } from './LabeledValueRow'
+import { cancelAllUploads, useUploadCounts } from '../stores/uploads'
+import { cancelAllDownloads, useDownloadCounts } from '../stores/downloads'
+import { Button } from './Button'
+import { RowGroup } from './Group'
+import { InputRow } from './InputRow'
+import { useInputValue } from '../hooks/useInputValue'
+import { setMaxUploads, useMaxUploads } from '../managers/uploadsPool'
+import { setMaxDownloads, useMaxDownloads } from '../managers/downloadsPool'
+
+export function SettingsAdvancedTransfers() {
+  const uploadCounts = useUploadCounts()
+  const downloadCounts = useDownloadCounts()
+  const maxUploads = useMaxUploads()
+  const maxDownloads = useMaxDownloads()
+
+  const maxUploadsInputProps = useInputValue({
+    value: String(maxUploads.data),
+    save: (text) => {
+      const n = Number(text.replace(/[^0-9]/g, ''))
+      if (Number.isFinite(n) && n > 0) setMaxUploads(n)
+    },
+  })
+
+  const maxDownloadsInputProps = useInputValue({
+    value: String(maxDownloads.data),
+    save: (text) => {
+      const n = Number(text.replace(/[^0-9]/g, ''))
+      if (Number.isFinite(n) && n > 0) setMaxDownloads(n)
+    },
+  })
+
+  return (
+    <>
+      <RowGroup title="Uploads">
+        <InfoCard>
+          <InputRow
+            label="Max concurrent uploads"
+            labelWidth={200}
+            keyboardType="number-pad"
+            {...maxUploadsInputProps}
+          />
+          <LabeledValueRow
+            label="Queued"
+            value={String(uploadCounts.totalQueued)}
+            canCopy={false}
+          />
+          <LabeledValueRow
+            label="Active"
+            value={String(uploadCounts.totalActive)}
+            canCopy={false}
+          />
+        </InfoCard>
+        <Button
+          style={{ marginTop: 10 }}
+          disabled={uploadCounts.total === 0}
+          onPress={() => {
+            cancelAllUploads()
+          }}
+        >
+          Cancel uploads
+        </Button>
+      </RowGroup>
+      <RowGroup title="Downloads">
+        <InfoCard>
+          <InputRow
+            label="Max concurrent downloads"
+            labelWidth={200}
+            keyboardType="number-pad"
+            {...maxDownloadsInputProps}
+          />
+          <LabeledValueRow
+            label="Queued"
+            value={String(downloadCounts.totalQueued)}
+            canCopy={false}
+          />
+          <LabeledValueRow
+            label="Active"
+            value={String(downloadCounts.totalActive)}
+            canCopy={false}
+          />
+        </InfoCard>
+        <Button
+          style={{ marginTop: 10 }}
+          disabled={downloadCounts.total === 0}
+          onPress={() => {
+            cancelAllDownloads()
+          }}
+        >
+          Cancel downloads
+        </Button>
+      </RowGroup>
+    </>
+  )
+}
